requirements.txt
asgiref==3.5.2
certifi==2023.7.22
charset-normalizer==3.3.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.2
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==2.0.7
